+++ src/components/atoms/NewChatFunction.jsx
